+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - The Handover/new project form collects the advisory board date and any
   conditions.
+- Users can be assigned to projects roles via 'change' links on the project
+  information page. Only users with the team lead role can assign users.
 
 ### Changed